backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
flask>=3.0.0
flask-cors
Pillow
scikit-learn
opencv-python
numpy